Add InitialDataBase Migration

+++ HR_Management.Domain/Entities/LeaveAllocation.cs
@@ -5,7 +5,7 @@
         #region Properties
 
         public int NumberOfDays { get; set; }
-        public int LeaveTypeId { get; set; }
+        public long LeaveTypeId { get; set; }
         public int Priod { get; set; }
 
         #endregion
