+++ HR_Management.Domain/Entities/LeaveRequest.cs
@@ -6,7 +6,7 @@
 
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
-        public int LeaveTypeId { get; set; }
+        public long LeaveTypeId { get; set; }
         public DateTime RequestDate { get; set; }
         public string RequestComments { get; set; }
         public DateTime? ActionDate { get; set; }
